Use Swagger to generate rest docs
Resulting documentation example is attached. We can change the styling
to other bootswatch templates. I've chosen one that is similar to
Alluxio colors (blue).

Docs are generated in `generated/<service>/index.html`. So far the job,
master, worker, and proxy are supported. Deprecated APIs are not
documented, and internal APIs (ie. for the web ui) are not documented
because we do not want others to rely on them.
![Screen Shot 2019-05-16 at 5 47 20
PM](https://user-images.githubusercontent.com/2223813/57895890-b6950d00-7802-11e9-852e-11213ba0a748.png)

pr-link: Alluxio/alluxio#9095
change-id: cid-b84badfe53cec0731879d17af964e7bb58cf61bb