Add extended attributes to inodes
This change will introduce a new field on our inode termed
`xAttr` which is an abbreviation for "Extended Attributes". This
is a `map<string, string>` structure can be utilized in various
ways to add more information to inodes past the standard set
of fields. It gives much more flexibility when adding new features
which may not be used on all inodes (i.e. pinning, TTL). This can
increase the space efficiency of inodes, and possibly decrease
heap usage if utilized only when necessary.

This change will not expose the `xattr` to clients yet, however it
could be in the future. For now the `xattr` kv-pairs are solely for
use between the FileSystemMaster and the Ufs Modules

pr-link: Alluxio/alluxio#9057
change-id: cid-cfb4c9a90fd64289619c43ec1f7ffc40e0d66d2c