Use fair lock.
Even if reinit acquires the write lock, it still needs to acquire the counter resource to allow reinit. If there are ongoing RPCs, reinit will not happen even if this write lock is acquired.

So the only point of having a better chance to lock the write lock here is:
when reinit is called, existing ongoing RPCs finish, but there are new RPCs calling acquireResourceToBlockReinit concurrently, in this case, the write lock might starve and reinit will never happen.

To solve the above problem, I prefer to change the lock to a fair lock, and just lock the write lock instead of tryLock.