Fix locality in containerized environments
In a containerized env such as k8s, the worker and client hostname may
not match. To determine if a worker is local, file system inspection is
used instead when property
`alluxio.worker.data.server.domain.socket.as.uuid=true` is set.

In this PR:
- update the method to choose a local worker for BlockInStream/OutStream
- if client container does not mount domain socket dir but has the same
hostname as worker, do not attempt to open local file short circuit
stream if worker supports domain socket (as /dev/shm is not accessible)
- unmount as part of fuse startup in case k8s shutdown hook is flaky

pr-link: Alluxio/alluxio#9143
change-id: cid-eeb9aa299b842ab7086bde957649332a68416162