Update LockCache's eviction behavior to avoid deadlock
Previously, when the soft limit is reached, and no lock can be evicted,
calls to get and tryGet will deadlock because the eviction thread will
be blocked in evictIfOverLimit because no lock can be evicted, while
other threads will block waiting for the eviction thread to finish.

This is a temporary fix for 2.0 release. The root cause is that in
master, some operations will try to hold locks for all inodes and edges
under a path recursively, and these locks will only be released when the
operation finishes. This causes a deadlock because if the number of
locks to be acquired is larger than the LockCache size, then no lock can
be acquired, so the operation is blocked, consequently, the existing
locks can not be released.

A better solution is to add an iterator interface to InodeTree to lock
and unlock paths while traversing.

pr-link: Alluxio/alluxio#9153
change-id: cid-babbb135edbd87fa022c288f450747a84353fb0b