Implement authentication long polling
`AuthenticationServer` has a setting for revoking authentication after a
period of inactivity. To handle that on the client side, metadata
clients, after a period of inactivity, will retry after getting
`Unauthenticated` code. However, due to nature of streaming, data
clients can not retry after getting the error because they might have
pipelined more data before seeing the error. And since this revocation
will not change the connection state, they used to continue getting
`Unauthenticated`. See https://github.com/Alluxio/alluxio/issues/9089
for an instance of this problem.

This PR introduces long polling to authentication handshake. Client and
server will not complete streams used for authentication and instead
will use it for notifying end of an authentication session. With this
change, revocation on server will be propagated to client channel via
health status, causing a client recreation for later use of the same
channel. Also client closing the channel will notify server and it'll
clean its state of the recently closed channel.

Periodic cleanup has not been disabled in order to not prolong a
duration for a channel to remain authenticated.

pr-link: Alluxio/alluxio#9149
change-id: cid-7c847b674046f4836c2d77881d47985a760b8951