Reduce the information allocator need in createBlockMeta
In each `createBlockMetaInternal`, Space allocator will get a updated
`BlockMetadataManagerView` which will do a full copy of pinList. Copy
the pinList is slow when the list is huge (especially in async through)
and copy is down in a `synchronized lock` which reduce the concurrency.
And actually, the allocator in the `createBlockMetaInternal` do not need
the pin list info. It only needs the storage tier and directory
information.

This PR is to reduce the information given to allocator and thus reduce
the contention and improve performance.

pr-link: Alluxio/alluxio#9335
change-id: cid-c2b0d42bd288ce558c99a101e1c52078e74ee44f