Fix version string in shaded-hadoop module
This is the quick fix for #9416 thanks to @asin929 in 2.0 branch.
This version string is correct to be `${ufs.hadoop.version}` in master
(2.1.0-SNAPSHOT) branch but was updated and hardcoded to be
`2.2.1-SNAPSHOT` in branch-2.0 due to maven release plugin.

Talked to @calvinjia , we should modify the POM to mark this module to
be skipped during release process.

pr-link: Alluxio/alluxio#9428
change-id: cid-177499ded30367de5c953045574e85c6b62da96e