Perform master-side check for the syncInterval option
Instead of blindly accepting the option value, this change will
now check the common options for `hasSyncIntervalMs` before
sending it to the locking scheme. If the value is not sent
from the client, then the fallback is to use the configured
value on the master.

This should prevent bugs on syncing too often if the sync
interval is left out on the client side for some reason.