Add metadataSyncInterval to client RPCs by default
The SetAttribute RPC in the client works via omission of
properties. If a property is included within the RPC options
object, then it is assumed the user set this property and
should then be set on the inode. This works for all fields
within the SetAttribute call except for the sync interval.

The metadata sync interval must be included on client RPCs.
Otherwise, if it is not included, then the
DefaultFileSystemMaster will assume that the user has
implicitly set a sync interval of 0 - which results in syncing
metadata for every RPC. Users can still override this value
from their configuration. However, by default it should be
sending a value of -1 (the default for the property key).

This change makes sure that the interval is set and adds a
small unit test.

Fixes #9470

pr-link: Alluxio/alluxio#9471
change-id: cid-3d8e1ae7d0a2cff12b9c34a32524ffc1f9238f2a