Add : space for exception message
**Description**
The exception message "Active Syncing is not supported on this UFS type"
+ ufsResource.get().getUnderFSType() will get the result "Active Syncing
is not supported on this UFS types3" for s3.

![image](https://user-images.githubusercontent.com/19620810/61782905-7d4bcf80-ae39-11e9-8aed-3e02f2df23ba.png)

**Improvement**
Append : space for the message: "Active Syncing is not supported on this
UFS type: ", And the result will be "Active Syncing is not supported on
this UFS type: s3" for s3.

pr-link: Alluxio/alluxio#9554
change-id: cid-3f96caf623e627e69aff27d3196a70e3692117e9