Fix Job master monitor when running in HA mode
The MasterHealthCheckClient uses a hard-coded fully-qualified
class name in order to do a ps -ef to check for the process on
another node.

The class name of the job master was previously updated, but the
enum with  the value here was not. This change fixes the class
name so that the health check can pass when running in HA
mode.

Fixes #9526

Initially reported by github user @lodocj