Validate only last line for leader command test
This is to deflake leader command integration test.

The reason for failure is that error output contains some netty error
trace that we can't control. When netty receives a command for a stream
that is closed or being closed, such traces are written to sys::err. In
this case, it's the latter, that some stream is about to be closed when
leader command hits the server. This flake is introduced by gRPC long
polling. Reason for that is now that servers keep an open stream,
closing a server attempts to close those as well. However that takes
some time after actual complete call is called on the stream. Hence it
increased exposure to receive cancelled/closed stream errors from netty.

pr-link: Alluxio/alluxio#9582
change-id: cid-c455b0b6604c60a490beed43b3a6f2e98424adf3