Refactor process launching in alluxio-start.sh
This change abstracts out the logic for launching Alluxio processes.
For most  users it is still recommended to use the alluxio-start.sh
script. The change here adds a new "alluxio-process" script which
is the launch point for any Alluxio server processes. When used
directly it launches the processes into the foreground of the
terminal using the bash built-in "exec". The reason why this is
needed is so that Alluxio can properly function under the
supervision of process supervisors such as SystemD, SysVInit,
and the userspace program supervisord.

Note that these changes preserve the function and arguments
of alluxio-start.sh in its entirety. This is an internal-logic refactor
which exposes the ability to start processes without using nohup.

The usage of alluxio-process is simple. Simply provide a process
name and the optional argument "-c" which will print the logging
output from the server to stdout.

In the future, the docker/K8s may be able to use these scripts instead
of the custom ones in the integration/docker folder. I am leaving that
as future work.

Closes #9389