Improve sync cache to consider recursive sync of ancestors
fixes #9545

This enables the ufs sync cache to inspect the ancestors of the path.
This allows the sync to be skipped if an ancestor was recently sync
recursively. This will potentially increase the time to check if sync
should be performed, since the path's parents have to be computed.

pr-link: Alluxio/alluxio#9668
change-id: cid-bd0cfe062c02c919585f5506bbe970d553dcc526