Improve ufs journal parsing efficiency
Fixes #9698

In the journal entry stream reader we keep a buffer to reuse for reading
entries. Because this buffer might not be fully utilized, we cannot use
protobuf's default `parseFrom(byte[])` and instead wrap the valid
portion of the buffer in a ByteArrayStream. However this causes
additional buffer allocations. We can instead use the lower level parser
API to parse a slice of the existing buffer.

This does not affect the embedded journal implementation because we
allocate a fresh buffer for each entry (exact size). Also I did not
apply this optimization for the v0 journals as those are no longer in
use aside for migrations.

The overall performance improvement is about 75% for a 10MB edit log
file stored on local disk, this also reduces the number of new objects
allocated which can reduce GC pressure.

pr-link: Alluxio/alluxio#9699
change-id: cid-5f1fdc9eb2cb39637dfb5518b0144839ab231827