Run k8s as root by default
After https://github.com/Alluxio/alluxio/pull/9363 is merged, we now run
the docker containers with user `alluxio` by default. `hostPath` volumes
in the default k8s specs are owned by `root` and the `alluxio` user does
not have access to write the journal to that location.

This PR allows choosing the user to run the Alluxio service on k8s in
the helm chart and replaces the generated template files.

pr-link: Alluxio/alluxio#9700
change-id: cid-32fd7427ad84f22cac7bf27717e1e1bf0138ba3b