Propagate thread-local AuthenticatedClientUser to gRPC executors
The proxy user information for impersonation is stored in the
ThreadLocal object in AuthenticatedClientUser class. The thread pools in
GrpcExecutors are not given this proxy user ThreadLocal when read/write
tasks are submitted to them. We have to write a wrapper for
ExecutorService that sets and unsets the ThreadLocal correctly.

pr-link: Alluxio/alluxio#9681
change-id: cid-c623177b4a388116e5fa209bd634b6da4161c847