Update ./bin/launch-process contains function so the script does not exit
The launch-process script uses the `set -e` flag which tells bash to exit
if a command returns non -zero. In this case, when the "contains"
function returned "1" (meaning that arg 1 contains arg 2 somewhere within
it), the script would exit.

This means that any users which set values through alluxio-env.sh could
possibly have their processes fail to start if setting things like Xmx
and MetaspaceSize that are checked in either the master or workers.

To remediate this issue the contains function was modified to
instead echo "1" or "0". Callers then simply have to call the
function in a subshell and use the result in their conditional
checks. This approach offers a better and more consistent usage
with the rest of our scripts.