Perform copycat connection building in a dedicated thread pool
Earlier, copycat catalyst `ThreadContext` was used to build connections
for transport clients/servers. There are some cluster states that are
managed by a single threaded `ThreadContext` and that could cause a node
to loose contact with rest of cluster while a connection is being built.
This PR offloads copycat connection building to a dedicated thread pool.

pr-link: Alluxio/alluxio#9738
change-id: cid-0a96f3d507106eb6d910848d673839e18c6530b1