Implement ShimFileSystem
`ShimFileSystem` is an alternative to existing `AlluxioFileSystem` where
foreign URIs are supported on client side. URIs that are sent to master
as-is need to be translated to internal Alluxio URIs. This is done by a
reverse lookup on master mount table. For `ShimFileSystem` to work
properly, the file system that it has been attached to must be mounted
on Alluxio.

Closes https://github.com/Alluxio/alluxio/issues/9231

pr-link: Alluxio/alluxio#9665
change-id: cid-2d51bfe5352e2ab4b824c0b49e06e26b8a0f3ec5