Avoid sync for getFileInfo if parent was recently listed (not recursive)
**Description**
Improving cache logic for PR #9668 #9714 . from #9715, #9771

**Improvement**
If the direct child of a recently listed parent is retrieved, the sync
is not necessary. This applies to `getFileInfo`, since that method does
not list the children. This PR allows the cache to avoid the sync if a
`getFileInfo` call is a direct child of a listed (not recursive) parent.

pr-link: Alluxio/alluxio#9773
change-id: cid-129d9375127f2e252e6f9b1716449151d1b8abaf