Make docker entrypoint handle signals properly
This PR improves the docker entrypoint script in a few ways:

- When launching single-process containers (i.e. worker-only, job worker,
master-only, proxy) the child process is now exec'd at the end of the
script so that the status of the container directly reflects the status
of the single alluxio process.
- Inside multi-process containers (master, worker) the child processes
will  now receive any signals which are sent to the parent shell
process.  This is achieved by using the `trap` feature of bash. to
forward the signals. See the comments for some more details

The main caveat with this implementation is that the signal
handler currently has a `wait` at the end of the handler. This disallows
the same signal to be sent multiple times. Without the  `wait` the
script will exit when the trap handler finishes. Not all signals
may cause Alluxio processes to exit, but without the wait, they will
forced to exit at any signal.

Resolves #9744