Use more specific Iceberg dependencies
Previously, the Hive UDB required all of iceberg's jars. This led to
some conflicts with older spark jars when developing integration
tests. By restricting the dependencies we get rid of conflicting
jars.

This change is a pre-requisite to full integration tests for the catalog
master. After this change we can use a local spark session to write out
parquet files which can then be consumed by an embedded HMS

Also, this PR increases the HMS version from 2.3.2 to 2.3.6

pr-link: Alluxio/alluxio#9888
change-id: cid-1b3019ec58f095018a0f01f4b8cb0f80eb1d7d11