Fix flakiness of migrate test
This test became flaky when the default write type became async-through.
This is because when the migrate is configured to delete the source,
sometimes, that inode is NOT persisted, so it is not removed from the
UFS. Then, on the `exists` call, it will loadMetadata from UFS, and it
will load the file from UFS and the deleted file will "return".

pr-link: Alluxio/alluxio#9884
change-id: cid-d9c66bd59b9e4180756dc13a8ac243f6951aafbe