Adding testing and logging to reproduce the issue
Run the FileOutStreamAsyncWriteIntegrationTest#badSpaceReserverTest

The test will trigger many jobs to run in the job service. It also sets
the max limit to only 500 jobs, but a retention of 0 seconds, so new jobs
coming in can run so long as there are jobs that have finished.

The issue we're seeing is that when we have many new jobs come in
Some will get scheduled, finish their tasks, and then be put into the
mFinishedJobs map with the time they finished. However, after extensive
testing I still see ResourceExhaustedExceptions being thrown, and when
they are thrown the state of the mIdToJobCoodinator map contains many
(hundreds) of jobs which say completed.

Logging has been added for the time when each job id gets added to
mFinished, removed from mFinished, and also removed from
mJobIdToCoordinator.

Essentially all the logs have shown me is that there are a number
of jobs which are created, run, finishes its task, gets added to
mFinished, but never logs a removal from the map.

The exception is only thrown when the map is empty. It seems that
we're having some kind of "ghost removal".