[WIP] Add JobTracker to JobMaster
The JobTracker encapsulates the functions that the mIdToCoordinator
ConcurrentHashMap and the mFinishedJobs ConcurrentSkipListMap. This
implementation has a few differences, namely switching out the
ConcurrentSkipListSet for a PriorityBlockingQueue instead.

From preliminary testing, the issue seems to be resolved. Mass amounts
of jobs getting added does not result in us hitting the limit. Further
testing is required. Additionally, the code is quite messy.