Fix bug for job info loss
mFinishedJobs in JobMaster is a ConcurrentSkipListSet to store finished
JobInfo. JobInfo implements the compareTo() function to compare the
JobInfos by mLastStatusChangeMs. For two different JobInfos, if the
mLastStatusChangeMs in is the same, there is only one JobInfo stored in
mFinishedJobs after the JobInfos added and one is lost.

mId in JobInfo is unique and can be used to compare JobInfos.

Signed-off-by: liuhongtong <hongtongliu@126.com>