Add git revision to ./bin/alluxio version
In light of preserving behavior, the command behaves the same as in
previous versions. This change adds the -r option which can be specified
optionally so that users can see part or all of the revision output
appended to the end of the normal version string.

Added two new maven plugins to retrieve the git revision.

Closes #9930

pr-link: Alluxio/alluxio#9936
change-id: cid-b4a1836b31b57881a1f9ae675dd53fe4d7534270