Fix local file writer client memory leak by cleaning up mmapped files
This PR addresses a memory leak in the Alluxio client for local file
writes on the short-circuit data access path. The impact is most
significant for long running processes, such as the job worker (with an
embedded Alluxio client), as open files are not released until the JVM
terminates and continue to occupy memory in the mean time.

In the issue reported, the worker and job worker run in the same
container. Pinning a 2G file causes set replication = 1 and the
replication checker creates a job for the job worker to load the file
from the ufs. The job worker opens mmaped files for temporary blocks and
does not cleanup temporary blocks correctly as reported by:
```
bash-4.4# lsof +L1 | grep /dev/shm
...
166 /usr/lib/jvm/java-1.8-openjdk/jre/bin/java
/dev/shm/alluxioworker/.tmp_blocks/4/109d8b9683aba404-1000014 (deleted)
166 /usr/lib/jvm/java-1.8-openjdk/jre/bin/java
/dev/shm/alluxioworker/.tmp_blocks/917/1c977804f85aa395-1000017
(deleted)
```

When running `du` and based on the worker metadata, there is space
available in `/dev/shm`. However when `df` is run, the job worker
process occupies memory for unclosed files which have been deleted.

Fixes: https://github.com/Alluxio/alluxio/issues/9666

pr-link: Alluxio/alluxio#10001
change-id: cid-f2a591f4202b96abbee286de1a7cf5fd89a18a1d