Remove from persist jobs synchronously
There is a case when `handleSuccess` is expensive (for example for
object stores). This leads to a large number of `handleSuccess` calls
being queued up, possibly conflicting with each other and consuming the
threadpool / queue.

I validated it with S3, but unfortunately it is nontrivial to make a
test for this case.

Fixes #9925

pr-link: Alluxio/alluxio#10015
change-id: cid-18f9fd3c412190c3c67d806cd77f84b31294bbc3