Add K8s Job dedicated for formatting master journal
This changes adds a set of Helm YAML templates.

The YAML creates a K8s Job that formats the master log and dies after
finishing. This provides a clean way for the user to format log without
going into the Pods, especially helpful when the masters are not started
or to clean up state without the need to start a master container.

pr-link: Alluxio/alluxio#10003
change-id: cid-5a20b7993aa3e61a9e2871b998fc5f31712fc5c9