Fix restore from backup on local path
There was a regression introduced that caused restoring from a
local backup file to stop working. Without this, all backups would
need to be stored in the root UFS which is non-ideal.

This PR fixes the regression and adds a test to ensure that we can
restore from backup when the root UFS is not local and the backup
path is on the local system.

- Regression originally introduced in #9808
- Continuation of PR #9972
- Fixes #9969

pr-link: Alluxio/alluxio#10046
change-id: cid-d158076c1e3d16238d72cf1bbc647ec639bf8c4e