Add block size to information stored in UfsFileStatus
Adds block size info to the `UfsFileStatus`. This mainly helps HDFS whos
implementation of `getBlockSizeBytes` has 2(!) rpc calls.

This change should be fully backward compatible, but older servers using
newer ufs jars will not work (because the constructor of UfsFileStatus
is different).

@apc999 A separate question is if we should enforce usage of UFS jars
which are on the same version of the server so we don't need to maintain
legacy code.

Fixes #10035

pr-link: Alluxio/alluxio#10057
change-id: cid-ad75817a4959cea2f0967a8792841ec05d1090c8