Write to the journal
It journals attachDB operation and addTable operation, mount operation
as part of the addTable is journaled by FileSystemMaster.
- Separated UdbTable and Table so that table will only get its state
from a udbTable when sync is called. This prevents the catalog service
from automatically refreshing the state from udb upon restart.
- Will add an update table journal entry when we have state changes
(transformation) in the future, or when sync gets called.

pr-link: Alluxio/alluxio#10071
change-id: cid-51bde05c3c904b3f7712daa84cb9c795a1e9eb32