Enable support for structured data
This is an initial implementation for handling structured data with
Alluxio. This introduces several components in Alluxio.
- TableMaster and TableClient
- UnderDatabase concept
- Transformations

This implementation restricts the possible scenarios to read use-cases
of existing hive tables.

#### TableMaster/Client
There is a new master type in Alluxio, which handles the Table concept.
It is primarily responsible for handling Table related metadata
operations, like schemas, to enable explicit support for structured
data. The client interacts with the master mostly for retrieving
metadata about the tables, and how to read the tables.

#### UnderDatabase
The UnderDatabase is a similar concept to the existing UnderFS concept,
but with structured data. This is an interface where different
implementations can access different database-like systems. Alluxio
TableMaster uses this interface to be able to access and manage metadata
from different systems.

#### Transformations
This implementation also introduces the concept of transformations of
tables. This allows Alluxio to transparently transform data (from the
udb), to a different representation of the data, which may improve data
access efficiency. The first examples of transformations include
converting some input formats (csv files) into parquet, which is
typically more efficient than other formats.

pr-link: Alluxio/alluxio#10000
change-id: cid-6f747134b11dafc24e534f3627162c3dcf4f0189