Make client throw NotFoundException when service is not found
Currently if a service is not found on server, the client will throw
`UnavailableException` which is confusing and causing unnecessary
retries. This change makes it throw `NotFoundException` which is what
the server is actually throwing. To differentiate it from other
`NotFoundException` such as file not found, a `ServiceNotFoundException`
is included as cause.

pr-link: Alluxio/alluxio#10174
change-id: cid-62087ee8ce242d29c440b08a3436e3407ac0d4ad